hikptool: Rectify the cleancode issue.
Rectify some cleancode alarms.Includes:
1. Add the function return value verification.
2. The logic for assigning values to output parameters of functions is optimized.
3. Use of parentheses.
4. Fix potential dead loop.
5. Fix unsigned-signed mixed use.

+++ net/nic/nic_fd/hikp_nic_fd.c
@@ -630,7 +630,7 @@ static int hikp_nic_query_fd_rules(struct hikp_cmd_header *req_header, const str
 	uint16_t max_key_bytes;
 	uint32_t left_buf_len;
 	size_t one_rule_size;
-	uint16_t idx;
+	uint32_t idx;
 	int ret = 0;
 
 	if (stage >= NIC_FD_STAGE_NUM) {
